refactor: put call back creation into dedicated functions
