+++ modules/cli/src/command/analysis/AnalysisCommand.ts
@@ -4,6 +4,12 @@ import {
   type Config,
   type ModuleMetadata,
   type ParsedAnalysisInput,
+  type ModuleListenerDatabaseInterface,
+  isModuleListenerDatabase,
+  type GenericReport,
+  type ModuleAnalysisInterface,
+  type ModuleListenerInterface,
+  type Result,
 } from "@fabernovel/heart-common"
 import { Command, InvalidArgumentError } from "commander"
 import type { PackageJsonModule } from "../../module/PackageJson.js"
@@ -15,6 +21,16 @@ import {
   createThresholdOption,
   type AnalysisOptions,
 } from "./AnalysisOption.js"
+import { initializeModules, loadEnvironmentVariables } from "../../module/ModuleLoader.js"
+import { migrateListenerDatabase } from "../../module/ModuleMigration.js"
+import { startAnalysis, notifyListenerModules } from "../../module/ModuleOrchestrator.js"
+
+type AnalysisSubcommandCallback = <C extends Config>(
+  verbose: boolean,
+  config: C,
+  threshold: number | undefined,
+  listenerModulesMetadataMapFiltered: Map<string, PackageJsonModule>
+) => Promise<void>
 
 function prepareOptionsForValidation(options: AnalysisOptions): ParsedAnalysisInput {
   return {
@@ -32,12 +48,7 @@ function prepareOptionsForValidation(options: AnalysisOptions): ParsedAnalysisIn
 export const createAnalysisSubcommand = <C extends Config>(
   moduleMetadata: ModuleMetadata,
   listenerModulesMetadataMap: Map<string, PackageJsonModule>,
-  callback: (
-    verbose: boolean,
-    config: C,
-    threshold: number | undefined,
-    listenerModulesMetadataMapFiltered: Map<string, PackageJsonModule>
-  ) => Promise<void>
+  callback: AnalysisSubcommandCallback
 ): Command => {
   const subcommand = new Command(moduleMetadata.id)
 
@@ -86,4 +97,52 @@ export const createAnalysisSubcommand = <C extends Config>(
     })
 
   return subcommand
+}
+
+/**
+ * Callback function called once the CLI command has been executed (the user hit "enter").
+ * @param config
+ * @param threshold
+ * @param listenerModulesMetadataMap Filtered map of listener modules metadata and their file path
+ */
+export function createAnalysisSubcommandCallback(
+  packageJsonModule: PackageJsonModule,
+  modulePath: string
+): AnalysisSubcommandCallback {
+  return async <C extends Config>(
+    verbose: boolean,
+    config: C,
+    threshold: number | undefined,
+    listenerModulesMetadataMap: Map<string, PackageJsonModule>
+  ) => {
+    // load and validate environment variables for the analysis and listeners modules
+    loadEnvironmentVariables(modulePath)
+    listenerModulesMetadataMap.forEach((_, listenerModulePath) => {
+      loadEnvironmentVariables(listenerModulePath)
+    })
+
+    // initialize the modules
+    const analysisModules = await initializeModules<ModuleAnalysisInterface<Config, GenericReport<Result>>>(
+      new Map([[modulePath, packageJsonModule]]),
+      verbose
+    )
+    const listenerModules = await initializeModules<ModuleListenerInterface>(
+      listenerModulesMetadataMap,
+      verbose
+    )
+
+    // run database migrations for listener database modules
+    const listenerDatabaseModules = listenerModules.filter(
+      (listenerModule): listenerModule is ModuleListenerDatabaseInterface =>
+        isModuleListenerDatabase(listenerModule)
+    )
+    if (listenerDatabaseModules.length > 0) {
+      await migrateListenerDatabase(listenerDatabaseModules)
+    }
+
+    const report = await startAnalysis(analysisModules[0], config, threshold)
+
+    // notify the listener modules
+    await notifyListenerModules(listenerModules, report)
+  }
 }
